chore: move clipboard files to shell/browser/api
clipboard no longer lives in the renderer, so move from shell/common/api to shell/browser/api
